requirements.txt
requests
upstox-python-sdk
sqlalchemy
psycopg2-binary
python-dotenv
flask
flask-cors
openai
pyyaml
slack-sdk
yfinance
pandas
numpy
scipy
jinja2
pydantic
python-dateutil
feedparser
